created a separate folder for shared styles such as styled links and main wrapper
