+++ src/components/Home.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from './common/Header';
+import Header from './Header';
 
 const Home = () => {
     return <div></div>;
