--- src/components/common/Main.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const MainWrapper = styled.section`
-    flex-grow: 1;
-`;
-
-const Main = () => {
-    return <MainWrapper></MainWrapper>;
-};
-
-export default Main;
